♻️ refactor: remove static LobeHub model cards
