--- packages/model-bank/src/aiModels/lobehub/chat/mapping.ts
@@ -1,17 +1,0 @@
-import type { AIChatModelCard } from '../../../types/aiModel';
-
-export const mappingChatModels: AIChatModelCard[] = [
-  {
-    abilities: {
-      functionCall: true,
-      reasoning: true,
-    },
-    enabled: true,
-    id: 'lobehub-onboarding-v1',
-    settings: {
-      extendParams: ['enableReasoning'],
-    },
-    type: 'chat',
-    visible: false,
-  },
-];
