--- packages/model-bank/src/aiModels/lobehub/embedding.ts
@@ -1,19 +1,0 @@
-import type { AIEmbeddingModelCard } from '../../types/aiModel';
-
-export const lobehubEmbeddingModels: AIEmbeddingModelCard[] = [
-  {
-    contextWindowTokens: 8192,
-    description:
-      'An efficient, cost-effective next-generation embedding model for retrieval and RAG scenarios.',
-    displayName: 'Text Embedding 3 Small',
-    enabled: true,
-    id: 'text-embedding-3-small',
-    maxDimension: 1536,
-    pricing: {
-      currency: 'USD',
-      units: [{ name: 'textInput', rate: 0.02, strategy: 'fixed', unit: 'millionTokens' }],
-    },
-    releasedAt: '2024-01-25',
-    type: 'embedding',
-  },
-];
