--- packages/model-bank/src/aiModels/lobehub/index.ts
@@ -1,15 +1,0 @@
-import { allModels } from './findModel';
-
-export { lobehubChatModels } from './chat';
-export { lobehubEmbeddingModels } from './embedding';
-export * from './findModel';
-export { lobehubImageModels } from './image';
-export * from './utils';
-export {
-  lobehubVideoModels,
-  seedance15ProParams,
-  seedance20FastParams,
-  seedance20Params,
-} from './video';
-
-export default allModels;
