--- packages/model-bank/src/aiModels/lobehub/utils.ts
@@ -1,104 +1,0 @@
-import type { ModelParamsSchema } from 'model-bank';
-
-// Common parameters for Imagen models
-export const imagenBaseParameters: ModelParamsSchema = {
-  aspectRatio: {
-    default: '1:1',
-    enum: ['1:1', '16:9', '9:16', '3:4', '4:3'],
-  },
-  prompt: { default: '' },
-};
-
-export const NANO_BANANA_ASPECT_RATIOS = [
-  'auto',
-  '1:1', // 1024x1024 / 2048x2048 / 4096x4096
-  '2:3', // 848x1264 / 1696x2528 / 3392x5056
-  '3:2', // 1264x848 / 2528x1696 / 5056x3392
-  '3:4', // 896x1200 / 1792x2400 / 3584x4800
-  '4:3', // 1200x896 / 2400x1792 / 4800x3584
-  '4:5', // 928x1152 / 1856x2304 / 3712x4608
-  '5:4', // 1152x928 / 2304x1856 / 4608x3712
-  '9:16', // 768x1376 / 1536x2752 / 3072x5504
-  '16:9', // 1376x768 / 2752x1536 / 5504x3072
-  '21:9', // 1584x672 / 3168x1344 / 6336x2688
-];
-
-export const nanoBananaParameters: ModelParamsSchema = {
-  aspectRatio: {
-    default: 'auto',
-    enum: NANO_BANANA_ASPECT_RATIOS,
-  },
-  imageUrls: {
-    default: [],
-  },
-  prompt: { default: '' },
-};
-
-export const nanoBananaProParameters: ModelParamsSchema = {
-  aspectRatio: {
-    default: 'auto',
-    enum: NANO_BANANA_ASPECT_RATIOS,
-  },
-  imageUrls: {
-    default: [],
-  },
-  prompt: { default: '' },
-  resolution: {
-    default: '1K',
-    enum: ['1K', '2K', '4K'],
-  },
-};
-
-export const NANO_BANANA_2_ASPECT_RATIOS = [
-  ...NANO_BANANA_ASPECT_RATIOS,
-  '1:4',
-  '4:1',
-  '1:8',
-  '8:1',
-];
-
-export const nanoBanana2Parameters: ModelParamsSchema = {
-  aspectRatio: {
-    default: 'auto',
-    enum: NANO_BANANA_2_ASPECT_RATIOS,
-  },
-  imageUrls: {
-    default: [],
-  },
-  prompt: { default: '' },
-  resolution: {
-    default: '1K',
-    enum: ['512', '1K', '2K', '4K'],
-  },
-};
-
-export const gptImage1Schema = {
-  imageUrls: { default: [], maxCount: 1, maxFileSize: 5 * 1024 * 1024 },
-  prompt: { default: '' },
-  size: {
-    default: 'auto',
-    enum: ['auto', '1024x1024', '1536x1024', '1024x1536'],
-  },
-};
-
-// gpt-image-2 accepts any resolution satisfying: max edge ≤ 3840px, both edges
-// multiples of 16px, aspect ratio ≤ 3:1, pixels between 655,360 and 8,294,400.
-// Until the schema/UI supports free-form W×H input, we expose the official
-// "Popular sizes" list from https://developers.openai.com/docs/guides/image-generation.
-export const gptImage2Schema = {
-  imageUrls: { default: [], maxCount: 1, maxFileSize: 5 * 1024 * 1024 },
-  prompt: { default: '' },
-  size: {
-    default: 'auto',
-    enum: [
-      'auto',
-      '1024x1024',
-      '1536x1024',
-      '1024x1536',
-      '2048x2048',
-      '2048x1152',
-      '3840x2160',
-      '2160x3840',
-    ],
-  },
-};
