+++ packages/model-bank/src/aiModels/openai.ts
@@ -7,7 +7,33 @@ import {
   type AITTSModelCard,
   type AIVideoModelCard,
 } from '../types/aiModel';
-import { gptImage1Schema, gptImage2Schema } from './lobehub';
+
+const gptImage1Schema = {
+  imageUrls: { default: [], maxCount: 1, maxFileSize: 5 * 1024 * 1024 },
+  prompt: { default: '' },
+  size: {
+    default: 'auto',
+    enum: ['auto', '1024x1024', '1536x1024', '1024x1536'],
+  },
+};
+
+const gptImage2Schema = {
+  imageUrls: { default: [], maxCount: 1, maxFileSize: 5 * 1024 * 1024 },
+  prompt: { default: '' },
+  size: {
+    default: 'auto',
+    enum: [
+      'auto',
+      '1024x1024',
+      '1536x1024',
+      '1024x1536',
+      '2048x2048',
+      '2048x1152',
+      '3840x2160',
+      '2160x3840',
+    ],
+  },
+};
 
 export const openaiChatModels: AIChatModelCard[] = [
   {
