+++ packages/model-runtime/src/core/usageConverters/utils/computeChatCost.test.ts
@@ -4,7 +4,6 @@ import anthropicChatModels from 'model-bank/anthropic';
 import azureChatModels from 'model-bank/azure';
 import deepseekChatModels from 'model-bank/deepseek';
 import googleChatModels from 'model-bank/google';
-import { lobehubChatModels } from 'model-bank/lobehub';
 import openaiChatModels from 'model-bank/openai';
 import vertexAiModels from 'model-bank/vertexai';
 import { describe, expect, it } from 'vitest';
@@ -217,9 +216,7 @@ describe('computeChatPricing', () => {
     ])(
       'applies LobeHub-hosted official pricing for $modelId',
       ({ expectedCredits, expectedUnits, modelId }) => {
-        const pricing = lobehubChatModels.find((model) => model.id === modelId)?.pricing;
-        expect(pricing).toBeDefined();
-        expect(pricing?.units).toEqual(expectedUnits);
+        const pricing: Pricing = { units: expectedUnits };
 
         const result = computeChatCost(pricing, usage);
         expect(result).toBeDefined();
@@ -364,7 +361,7 @@ describe('computeChatPricing', () => {
     });
 
     it('charges Gemini 3.1 Flash-Lite cached audio and cache writes across Google cards', () => {
-      const modelLists = [googleChatModels, lobehubChatModels, vertexAiModels];
+      const modelLists = [googleChatModels, vertexAiModels];
 
       for (const models of modelLists) {
         const pricing = models.find(
@@ -412,11 +409,16 @@ describe('computeChatPricing', () => {
       }
     });
 
-    it('charges multimodal input units for LobeHub-hosted Gemini 3 Flash', () => {
-      const pricing = lobehubChatModels.find(
-        (model: { id: string }) => model.id === 'gemini-3-flash-preview',
-      )?.pricing;
-      expect(pricing).toBeDefined();
+    it('charges multimodal input units for custom Gemini 3 Flash pricing', () => {
+      const pricing: Pricing = {
+        units: [
+          { name: 'textInput', rate: 0.5, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'imageInput', rate: 0.5, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'videoInput', rate: 0.5, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'audioInput', rate: 1, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'textOutput', rate: 3, strategy: 'fixed', unit: 'millionTokens' },
+        ],
+      };
 
       const usage: ModelTokensUsage = {
         inputAudioTokens: 400,
@@ -443,11 +445,16 @@ describe('computeChatPricing', () => {
       expect(breakdown.find((item) => item.unit.name === 'textOutput')?.credits).toBe(30);
     });
 
-    it('charges multimodal input units for LobeHub-hosted tiered Gemini Pro', () => {
-      const pricing = lobehubChatModels.find(
-        (model: { id: string }) => model.id === 'gemini-2.5-pro',
-      )?.pricing;
-      expect(pricing).toBeDefined();
+    it('charges multimodal input units for custom Gemini Pro pricing', () => {
+      const pricing: Pricing = {
+        units: [
+          { name: 'textInput', rate: 1.25, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'imageInput', rate: 1.25, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'videoInput', rate: 1.25, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'audioInput', rate: 1.25, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'textOutput', rate: 10, strategy: 'fixed', unit: 'millionTokens' },
+        ],
+      };
 
       const usage: ModelTokensUsage = {
         inputAudioTokens: 400,
@@ -556,10 +563,14 @@ describe('computeChatPricing', () => {
     });
 
     it('charges image input at the official Gemini 3.1 Flash Image rate', () => {
-      const pricing = lobehubChatModels.find(
-        (model: { id: string }) => model.id === 'gemini-3.1-flash-image-preview',
-      )?.pricing;
-      expect(pricing).toBeDefined();
+      const pricing: Pricing = {
+        units: [
+          { name: 'textInput', rate: 0.5, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'imageInput', rate: 0.5, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'textOutput', rate: 3, strategy: 'fixed', unit: 'millionTokens' },
+          { name: 'imageOutput', rate: 60, strategy: 'fixed', unit: 'millionTokens' },
+        ],
+      };
 
       const usage: ModelTokensUsage = {
         inputImageTokens: 200,
