+++ src/store/image/slices/generationConfig/initialState.ts
@@ -1,11 +1,41 @@
 import { type ModelParamsSchema, type RuntimeImageGenParams } from 'model-bank';
 import { extractDefaultValues, ModelProvider } from 'model-bank';
-import { nanoBanana2Parameters } from 'model-bank/lobehub';
 
 import { DEFAULT_IMAGE_CONFIG } from '@/const/settings';
 
 export const DEFAULT_AI_IMAGE_PROVIDER = ModelProvider.Google;
 export const DEFAULT_AI_IMAGE_MODEL = 'gemini-3.1-flash-image-preview:image';
+
+const nanoBanana2Parameters: ModelParamsSchema = {
+  aspectRatio: {
+    default: 'auto',
+    enum: [
+      'auto',
+      '1:1',
+      '2:3',
+      '3:2',
+      '3:4',
+      '4:3',
+      '4:5',
+      '5:4',
+      '9:16',
+      '16:9',
+      '21:9',
+      '1:4',
+      '4:1',
+      '1:8',
+      '8:1',
+    ],
+  },
+  imageUrls: {
+    default: [],
+  },
+  prompt: { default: '' },
+  resolution: {
+    default: '1K',
+    enum: ['512', '1K', '2K', '4K'],
+  },
+};
 
 export interface GenerationConfigState {
   parameters: RuntimeImageGenParams;
